+++ app/src/main/java/com/github/openwebnet/model/IpcamModel.java
@@ -120,7 +120,7 @@ public class IpcamModel extends RealmObject
 
         public Builder(Map<String, Object> map) {
             this.uuid = (String) map.get(FIELD_UUID);
-            this.environmentId = (Integer) map.get(FIELD_ENVIRONMENT_ID);
+            this.environmentId = ((Long) map.get(FIELD_ENVIRONMENT_ID)).intValue();
             this.name = (String) map.get(FIELD_NAME);
             this.url = (String) map.get(FIELD_URL);
             this.type =(String) map.get(FIELD_STREAM_TYPE);
